+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 2.2.13
+
+### Patch Changes
+
+- @latticexyz/schema-type@2.2.13
+
 ## 2.2.12
 
 ### Patch Changes