+++ packages/world-module-metadata/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @latticexyz/world-module-metadata
 
+## 2.2.13
+
+### Patch Changes
+
+- @latticexyz/schema-type@2.2.13
+- @latticexyz/store@2.2.13
+- @latticexyz/world@2.2.13
+
 ## 2.2.12
 
 ### Patch Changes